переводим файлы js на модульность
